refactor(Home): refact Home structure
