+++ src/pages/home/Home.tsx
@@ -1,7 +1,6 @@
-
-import Footer from '@/components/Footer/Footer';
-import MiddleContent from './components/MiddleContent/MiddleContent';
-import HomeSearch from './components/HomeSearch/HomeSearch';
+import Footer from "@/components/Footer/Footer";
+import MiddleContent from ".//MiddleContent/MiddleContent";
+import HomeSearch from ".//HomeSearch/HomeSearch";
 
 interface HomeProps {
   imgMapa?: string;
@@ -16,7 +15,6 @@ export default function Home({ imgMapa, nm_pedflex }: HomeProps) {
       <MiddleContent imgMapa={imgMapa} nm_pedflex={nm_pedflex} />
 
       <Footer />
-
     </>
   );
 }
