+++ src/pages/home/HomeSearch/HomeSearch.styles.ts
