+++ src/pages/home/HomeSearch/HomeSearch.tsx
@@ -1,18 +1,18 @@
-import { useTheme } from 'styled-components';
+import { useTheme } from "styled-components";
 import {
   DescriptionContent,
   HomeContainer,
   HomeContent,
   LogoContent,
   LogoImg,
-} from './HomeSearch.styles';
-import { HomePageButton } from '../../../../components/HomeButton/HomeButton';
-import React from 'react';
-import { cpfMask, cnpjMask } from '@/lib/masks';
-import HomeSearchInput from '@/components/HomeSearchInput/HomeSearchInput';
+} from "./HomeSearch.styles";
+import { HomePageButton } from "@/components/HomeButton/HomeButton";
+import React from "react";
+import { cpfMask, cnpjMask } from "@/utils/masks";
+import HomeSearchInput from "@/components/HomeSearchInput/HomeSearchInput";
 
 export default function HomeSearch() {
-  const [cnpj, setCnpj] = React.useState('');
+  const [cnpj, setCnpj] = React.useState("");
 
   const theme = useTheme();
 
@@ -31,14 +31,14 @@ export default function HomeSearch() {
 
       <HomeContent>
         <DescriptionContent>
-          Uma seleção de produtos com a cara da sua loja a um{' '}
+          Uma seleção de produtos com a cara da sua loja a um{" "}
           <span style={{ color: theme.colors.primary, fontWeight: 600 }}>
             click
-          </span>{' '}
+          </span>{" "}
           de distância
         </DescriptionContent>
 
-        <div style={{ width: '80%' }}>
+        <div style={{ width: "80%" }}>
           <HomeSearchInput
             placeholder="Informe seu CPF/CNPJ para ver as ofertas"
             value={maskedInput(cnpj)}
@@ -50,7 +50,7 @@ export default function HomeSearch() {
           size="lg"
           title="Buscar"
           icon="assets/lupa.png"
-          style={{ height: '60px', width: '200px' }}
+          style={{ height: "60px", width: "200px" }}
         />
       </HomeContent>
     </HomeContainer>
