+++ src/pages/home/MiddleContent/MiddleContent.styles.ts
