+++ src/pages/home/MiddleContent/MiddleContent.tsx
